chatbot/custom/requirements.txt
langchain 
langchain-community 
groq 
faiss-cpu 
ollama 
firecrawl-py
openai
python-dotenv
langchain_ollama